Overhaul recipe actions to be in editor
