+++ websites/recipe-website/editor/controller/actions/update.ts
@@ -2,10 +2,10 @@
 
 import parseRecipeFormData from "../parseFormData";
 import { RecipeFormState } from "../formState";
-import { Recipe } from "../types";
-import createDefaultSlug from "../createSlug";
+import { Recipe } from "recipe-website-common/controller/types";
+import createDefaultSlug from "recipe-website-common/controller/createSlug";
 import slugify from "@sindresorhus/slugify";
-import getRecipeBySlug from "../data/read";
+import getRecipeBySlug from "recipe-website-common/controller/data/read";
 import { commitContentChanges } from "content-engine/git/commit";
 import {
   handleSuccess,
